requirements.txt
gradio>=5.0.0,<6
huggingface_hub>=0.24.0
transformers>=4.44.2
sentence-transformers>=3.0.1
numpy>=1.26.4
pandas>=2.2.2
pydantic>=2.8.2
requests>=2.32.3
duckduckgo-search>=6.2.9
beautifulsoup4>=4.12.3
urllib3>=2.2.2
charset-normalizer>=3.3.2
python-dotenv>=1.0.1
